Add support for Google App Engine to the SemiDBM. This commit will allow GAE users to create a semidbm database offline, then upload it to GAE, and then use that database in read-only mode within GAE.
Google App Engine runs python in a sandbox, and does not support os file descriptor access that is used by semidbm. It also does not support opening files for writing (but it does support read only).

This commit changes those function calls to the standard python open, read, and write functions, which are supported in the GAE sandbox.  Further, if the semidbm database is opened in read-only mode, then the appropriate flags are used which work on GAE.